# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AbacusAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/student/session/start:
    post:
      operationId: startSession
      tags: [student]
      summary: Start a practice session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStartResponse"
        "403":
          description: Plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeResponse"
  /v1/student/session/complete:
    post:
      operationId: completeSession
      tags: [student]
      summary: Complete a practice session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteSessionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResult"
  /v1/student/sessions:
    get:
      operationId: listSessions
      tags: [student]
      summary: List recent sessions
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
...
  /v1/ai/generate-problems:
    post:
      operationId: generateProblems
      tags: [ai]
      summary: Generate AI practice problems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateProblemsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Problem"
        "403":
          description: Plan required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeResponse"
...
  /v1/parent/child/{childId}/progress:
    get:
      operationId: getChildProgress
      tags: [parent]
      summary: Get child progress
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"
...
components:
  schemas:
...
    UpgradeResponse:
      type: object
      properties:
        upgrade:
          type: boolean
        message:
          type: string
        requiredPlan:
          type: string
      required:
        - upgrade
        - message
        - requiredPlan
...
    StudentProgress:
      type: object
      properties:
        dailyAccuracy:
          type: array
          items:
            $ref: "#/components/schemas/DailyAccuracy"
        levelProgress:
          type: array
          items:
            $ref: "#/components/schemas/LevelProgress"
        totalSessions:
          type: integer
        avgAccuracy:
          type: number
        avgDuration:
          type: number
        hintsUsedTotal:
          type: integer
        strongAreas:
          type: array
          items:
            type: string
        weakAreas:
          type: array
          items:
            type: string
      required:
        - dailyAccuracy
        - levelProgress
        - totalSessions
        - avgAccuracy
        - avgDuration
        - hintsUsedTotal
        - strongAreas
        - weakAreas
    DailyAccuracy:
      type: object
      properties:
        date:
          type: string
        accuracy:
          type: number
        sessions:
          type: integer
      required:
        - date
        - accuracy
        - sessions
    LevelProgress:
      type: object
      properties:
        level:
          type: integer
        sessionsCompleted:
          type: integer
        avgAccuracy:
          type: number
      required:
        - level
        - sessionsCompleted
        - avgAccuracy
    StartSessionBody:
      type: object
      properties:
        level:
          type: integer
      required:
        - level
    SessionStartResponse:
      type: object
      properties:
        sessionId:
          type: string
        problems:
          type: array
          items:
            $ref: "#/components/schemas/Problem"
        level:
          type: integer
        difficulty:
          type: string
      required:
        - sessionId
        - problems
        - level
        - difficulty
    CompleteSessionBody:
      type: object
      properties:
        sessionId:
          type: string
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ProblemResult"
        duration:
          type: integer
        hintsUsed:
          type: integer
      required:
        - sessionId
        - problems
        - duration
        - hintsUsed
    SessionResult:
      type: object
      properties:
        score:
          type: integer
        accuracy:
          type: number
        xpEarned:
          type: integer
        badges:
          type: array
          items:
            type: string
        newLevel:
          type: ["integer", "null"]
        aiAdapt:
          $ref: "#/components/schemas/AdaptResponse"
        streakUpdated:
          type: boolean
      required:
        - score
        - accuracy
        - xpEarned
        - badges
        - streakUpdated
    Session:
      type: object
      properties:
        id:
          type: string
        level:
          type: integer
        score:
          type: integer
        accuracy:
          type: number
        duration:
          type: integer
        hintsUsed:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - level
        - score
        - accuracy
        - duration
        - hintsUsed
        - createdAt
    Problem:
      type: object
      properties:
        question:
          type: string
        answer:
          type: number
        hint:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
      required:
        - question
        - answer
        - hint
        - difficulty
    ProblemResult:
      type: object
      properties:
        question:
          type: string
        answer:
          type: number
        studentAnswer:
          type: ["number", "null"]
        correct:
          type: boolean
      required:
        - question
        - answer
        - correct
...
    GenerateProblemsBody:
      type: object
      properties:
        level:
          type: integer
        recentAccuracy:
          type: number
        focusArea:
          type: string
        studentAge:
          type: integer
      required:
        - level
        - recentAccuracy
        - focusArea
        - studentAge
...
    AdaptResponse:
      type: object
      properties:
        next_difficulty:
          type: string
          enum: [easy, medium, hard]
        recommended_problem_count:
          type: integer
        focus_area:
          type: string
        should_advance_level:
          type: boolean
        coach_tip:
          type: string
      required:
        - next_difficulty
        - recommended_problem_count
        - focus_area
        - should_advance_level
        - coach_tip
...